pyproject.toml
[tool.poetry]
name = "mtg_pl"
version = "2020.3.1"
description = "MTG po polsku"
authors = ["Dominik Kozaczko <dominik@kozaczko.info>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
requests = "^2.23.0"
sphinx-rtd-theme = "^0.4.3"
polib = "^1.1.0"
python-decouple = "^3.4"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.1.1"
sphinx = "^2.4.3"
ipython = "^7.20.0"
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry>=1.1"]
build-backend = "poetry.masonry.api"
